requirements.txt
stomp.py==4.1.0